requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
django==5.0.6
    # via -r requirements.in
idna==3.7
    # via requests
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
urllib3==2.2.2
    # via requests
